+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - Added the Elevated Session Duration test.
 - Added the Web Alias In Base URL test.
 - Added PHP version support thresholds up until PHP 8.0 ([supported versions](https://www.php.net/supported-versions.php)).
+- Added integration with Sentry.io. 
 - Added logging to a dedicated `sherlock.log` file.
 
 ### Changed